refactor(sb): Fix sb server main list protocol

+++ src/Servers/QueryReport/src/Abstraction/BaseClass/ResponseBase.cs
@@ -8,8 +8,11 @@ namespace UniSpyServer.Servers.QueryReport.Abstraction.BaseClass
     {
         protected new RequestBase _request => (RequestBase)base._request;
         protected new ResultBase _result => (ResultBase)base._result;
-        public new byte[] SendingBuffer{ get => (byte[])base.SendingBuffer;
-            protected set => base.SendingBuffer = value; }
+        public new byte[] SendingBuffer
+        {
+            get => (byte[])base.SendingBuffer;
+            protected set => base.SendingBuffer = value;
+        }
         protected ResponseBase(UniSpyRequestBase request, UniSpyResultBase result) : base(request, result)
         {
         }
@@ -19,7 +22,7 @@ namespace UniSpyServer.Servers.QueryReport.Abstraction.BaseClass
             List<byte> data = new List<byte>();
             data.AddRange(RequestBase.MagicData);
             data.Add((byte)_request.CommandName);
-            data.AddRange(BitConverter.GetBytes(_request.InstantKey));
+            data.AddRange(BitConverter.GetBytes((uint)_request.InstantKey));
             SendingBuffer = data.ToArray();
         }
     }
